Add tests for Partition. Also, format and lint.

+++ src/lib/storage/partitioning/round_robin_partition_schema.cpp
@@ -16,14 +16,13 @@ void RoundRobinPartitionSchema::append(std::vector<AllTypeVariant> values) {
   if (_next_partition >= static_cast<PartitionID>(_partitions.size())) {
     _next_partition = 0;
   } else {
-    _next_partition = static_cast<PartitionID>(static_cast<uint16_t>(_next_partition) + (uint16_t) 1);
+    _next_partition = static_cast<PartitionID>(_next_partition + 1);
   }
 }
 
 PartitionID RoundRobinPartitionSchema::get_matching_partition_for(std::vector<AllTypeVariant> values) {
-  _next_partition = static_cast<PartitionID>(static_cast<uint16_t>(_next_partition) + (uint16_t) 1);
+  _next_partition = static_cast<PartitionID>(_next_partition + 1);
   return _next_partition;
 }
-
 
 }  // namespace opossum
